RP2350 : Remove the old filter functions that used double precision
Switch to the single precision biquad implementation

+++ firmware/DSPi/config.h
@@ -307,7 +307,7 @@ typedef struct __attribute__((packed)) {
 typedef struct {
     // biquad filter
     float b0, b1, b2, a1, a2;    // float biquad coefficients
-    double s1, s2;               // double precision state accumulators (Using inline DCP)
+    float s1, s2;                // single precision state accumulators
 
     // state variable filter
     float sva1, sva2, sva3, svm0, svm1, svm2; // state variable filter coefficients
